VotingController: mulitple fixes+optiminzations

+++ src/libraries/Events.sol
@@ -21,18 +21,18 @@ library Events {
     event PoolRemoved(bytes32 indexed poolId);
 
     // vote(), migrateVotes()
-    event Voted(uint256 indexed epoch, address indexed caller, bytes32[] poolIds, uint256[] votes, bool isDelegated);
-    event VotesMigrated(uint256 indexed epoch, address indexed caller, bytes32[] srcPoolIds, bytes32[] dstPoolIds, uint256[] votes, bool isDelegated);
+    event Voted(uint256 indexed epoch, address indexed caller, bytes32[] poolIds, uint128[] votes, bool isDelegated);
+    event VotesMigrated(uint256 indexed epoch, address indexed caller, bytes32[] srcPoolIds, bytes32[] dstPoolIds, uint128[] votes, bool isDelegated);
     // delegate
     event DelegateRegistered(address indexed delegate, uint256 feePct);
     event DelegateFeeDecreased(address indexed delegate, uint256 currentFeePct, uint256 feePct);
     event DelegateFeeIncreased(address indexed delegate, uint256 currentFeePct, uint256 feePct, uint256 nextFeePctEpoch);
     event ClaimDelegateFees(address indexed delegate, uint256 feesClaimed);
-    // claimRewards
+    // voterClaimRewards
     event RewardsClaimed(address indexed caller, uint256 epoch, bytes32[] poolIds, uint256 totalClaimableRewards);
+    // claimDelegateFees
     event RewardsClaimedFromDelegate(uint256 indexed epoch, address indexed caller, address indexed delegate, bytes32[] poolIds, uint256 totalClaimableRewards);
-    
-    // delegatorsClaimRewardsFromDelegates/claimRewardsFromDelegate
+    // claimRewardsFromDelegate
     event RewardsClaimedFromDelegateBatch(uint256 indexed epoch, address indexed caller, address[] delegateList, bytes32[][] poolIdsPerDelegate, uint256[] totalUserNetRewards);
     event DelegateFeesClaimed(address indexed delegate, uint256 feesClaimed);
     // delegateClaimFeesFromDelegators
@@ -40,6 +40,7 @@ library Events {
 
     // claimSubsidies
     event SubsidiesClaimed(address indexed verifier, uint256 epoch, bytes32[] poolIds, uint256 totalSubsidiesClaimed);
+
 
     // depositSubsidies
     event SubsidiesDeposited(address indexed depositor, uint256 epoch, uint256 totalSubsidies);
